requirements.txt
flask
requests
colorama
websockets
python-dotenv
certifi
